Revert "Switch to ts"
