--- FirstTask/my-react-app/src/declarations.d.ts
@@ -1,4 +1,0 @@
-declare module "*.png" {
-  const value: string;
-  export default value;
-}
