--- FirstTask/my-react-app/src/shared/components/Interface/popupInterface.tsx
@@ -1,6 +1,0 @@
-import User from "./userInterface";
-export default interface PopupProps {
-  isOpen: Boolean;
-  onClose: () => void;
-  data: User | null;
-}
