--- FirstTask/my-react-app/src/shared/components/Interface/popupStateInterface.tsx
@@ -1,6 +1,0 @@
-import User from "./userInterface";
-export default interface PopupState {
-    isOpen: Boolean;
-    selectedUser: User | null;
-  }
-
