--- FirstTask/my-react-app/src/shared/components/Interface/userInterface.tsx
@@ -1,9 +1,0 @@
-export default interface User {
-  name: string;
-  phone: string;
-  email: string;
-  address: string;
-  position_name: string;
-  department: string;
-  hire_date: string;
-}
